update aMobile.vs
Added a getDevice function, that returns if the device is an iOS device, or an Android.
Added a function that will vibrate the device if used in tandem with Monaca.io (if the developer has the vibrate plugin enabled). This will also use the Web Vibration API as a fallback (if that browser supports it).
Made the library store the `isMobile` value as well as the client. Either can be referenced this way.